SHARED: inlined moveDelta

+++ src/modules/frontend/PlayerMovement.cpp
@@ -61,7 +61,7 @@ void PlayerMovement::update(float deltaFrameSeconds, float orientation, ClientEn
 	} else {
 		_move &= ~network::MoveDirection::JUMP;
 	}
-	const glm::vec3& newPos = Super::update(deltaFrameSeconds, orientation, speed, currentPos, heightResolver);
+	const glm::vec3& newPos = Super::update(deltaFrameSeconds, orientation, (float)speed, currentPos, heightResolver);
 	entity->setPosition(newPos);
 	entity->setAnimation(animation());
 }
